helpers: fold the physical-name tests into helpers_test.go
